Add CentOS support for role containerd
