+++ molecule/delegated/prepare/containerd.yml
@@ -3,3 +3,10 @@
   become: true
   ansible.builtin.apt:
     update_cache: true
+  when: "'Debian' in ansible_os_family"
+
+- name: Update package cache
+  become: true
+  ansible.builtin.dnf:
+    update_cache: true
+  when: "'RedHat' in ansible_os_family"
