+++ molecule/delegated/tests/containerd/main.py
@@ -1,6 +1,4 @@
-import pytest
-
-from .util.util import get_ansible, get_variable, get_from_url
+from ..util.util import get_ansible, get_variable
 
 testinfra_runner, testinfra_hosts = get_ansible()
 
@@ -32,26 +30,6 @@ def test_failpkg(host):
     for package_name in package_names:
         package = host.package(package_name)
         assert not package.is_installed
-
-
-def test_repo(host):
-    docker_configure_repository = get_variable(host, "docker_configure_repository")
-
-    if not docker_configure_repository:
-        pytest.skip("docker_configure_repository is not true")
-
-    package = host.package("apt-transport-https")
-    assert package.is_installed
-
-    key_content = get_from_url(get_variable(host, "docker_debian_repository_key"))
-
-    f = host.file("/etc/apt/trusted.gpg.d/docker.asc")
-    assert f.exists
-    assert not f.is_directory
-    assert f.mode == 0o644
-    assert f.user == "root"
-    assert f.group == "root"
-    assert f.content_string == key_content
 
 
 def test_pkg(host):
