+++ roles/containerd/defaults/main.yml
@@ -13,3 +13,6 @@ docker_configure_repository: true
 
 docker_debian_repository_key: https://download.docker.com/linux/ubuntu/gpg
 docker_debian_repository: "deb https://download.docker.com/linux/ubuntu {{ ansible_distribution_release }} stable"
+
+docker_redhat_repository_key: https://download.docker.com/linux/centos/gpg
+docker_redhat_repository: "https://download.docker.com/linux/centos/docker-ce.repo"
